Move FallbackFrom as an inner class of Fallback
