+++ src/main/java/org/cactoos/Fallback.java
@@ -23,6 +23,11 @@
  */
 package org.cactoos;
 
+import org.cactoos.iterable.IterableOf;
+import org.cactoos.iterable.Mapped;
+import org.cactoos.scalar.InheritanceLevel;
+import org.cactoos.scalar.MinOf;
+
 /**
  * Fallback from a {@link Throwable}.
  *
@@ -40,4 +45,63 @@ public interface Fallback<X> extends Func<Throwable, X> {
      *  is supported and {@link Integer#MIN_VALUE} otherwise
      */
     int support(Throwable exception);
+    
+    /**
+     * Fallback from exception.
+     *
+     * <p>There is no thread-safety guarantee.
+     *
+     * @param <T> Type of result
+     * @since 1.0
+     */
+    final class From<T> implements Fallback<T> {
+
+        /**
+         * The list of exceptions supported by this instance.
+         */
+        private final Iterable<Class<? extends Throwable>> exceptions;
+
+        /**
+         * Function that converts exceptions to the required type.
+         */
+        private final Func<Throwable, T> func;
+
+        /**
+         * Ctor.
+         * @param exp Supported exception type
+         * @param func Function that converts the given exception into required one
+         */
+        @SuppressWarnings("unchecked")
+        public From(final Class<? extends Throwable> exp,
+            final Func<Throwable, T> func) {
+            this(new IterableOf<>(exp), func);
+        }
+
+        /**
+         * Ctor.
+         * @param exps Supported exceptions types
+         * @param func Function that converts the given exception into required one
+         */
+        public From(
+            final Iterable<Class<? extends Throwable>> exps,
+            final Func<Throwable, T> func) {
+            this.exceptions = exps;
+            this.func = func;
+        }
+
+        @Override
+        public T apply(final Throwable exp) throws Exception {
+            return this.func.apply(exp);
+        }
+
+        @Override
+        public int support(final Throwable exception) {
+            return new MinOf(
+                new Mapped<>(
+                    supported -> new InheritanceLevel(exception.getClass(), supported).value(),
+                    this.exceptions
+                )
+            ).intValue();
+        }
+    }
 }
