+++ src/test/java/org/cactoos/func/FallbackFromTest.java
@@ -25,13 +25,14 @@ package org.cactoos.func;
 
 import java.io.IOException;
 import java.util.IllegalFormatWidthException;
-import org.cactoos.iterable.IterableOf;
+
+import org.cactoos.Fallback;
 import org.hamcrest.core.IsEqual;
 import org.junit.jupiter.api.Test;
 import org.llorllale.cactoos.matchers.Assertion;
 
 /**
- * Test case for {@link FallbackFrom}.
+ * Test case for {@link Fallback.From}.
  *
  * @since 0.31
  * @checkstyle JavadocMethodCheck (500 lines)
@@ -44,8 +45,8 @@ final class FallbackFromTest {
     void supportsException() {
         new Assertion<>(
             "Must support exactly exception class",
-            new FallbackFrom<>(
-                new IterableOf<>(IOException.class),
+            new Fallback.From<>(
+                IOException.class,
                 exp -> "IOException fallback"
             ).support(new IOException()),
             new IsEqual<>(0)
@@ -56,8 +57,8 @@ final class FallbackFromTest {
     void supportsInheritedException() {
         new Assertion<>(
             "Must support inherited exception class",
-            new FallbackFrom<>(
-                new IterableOf<>(RuntimeException.class),
+            new Fallback.From<>(
+                RuntimeException.class,
                 exp -> "RuntimeException fallback #1"
             ).support(new IllegalFormatWidthException(1)),
             new IsEqual<>(3)
@@ -68,8 +69,8 @@ final class FallbackFromTest {
     void doesNotSupportException() {
         new Assertion<>(
             "Must not support unrelated exception class",
-            new FallbackFrom<>(
-                new IterableOf<>(RuntimeException.class),
+            new Fallback.From<>(
+                RuntimeException.class,
                 exp -> "RuntimeException fallback #2"
             ).support(new ClassNotFoundException()),
             new IsEqual<>(Integer.MIN_VALUE)
